% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_nearest_mode2D}
\alias{find_nearest_mode2D}
\title{Find the nearest maximum in a matrix}
\usage{
find_nearest_mode2D(A, px, py)
}
\arguments{
\item{A}{A matrix.}

\item{px, py}{A point within the dimensions of the matrix.}
}
\value{
Returns the x and y matrix positions of the nearest maximum that was found.
}
\description{
Find the nearest maximum in a matrix to a given point by hillclimbing.
}
\details{
A hillclimbing search from the input point across the array to find the nearest local maximum. 
The search is through each square surrounding the next input point in the search (accounting for the array boundaries).
Ties are broken by taking the nearest euclidean distance (i.e. point which are vertical or horizontal from the input rather than diagonal)
and then random sampling. To make this reproducible the seed should be set.
}
\examples{
set.seed(444)
msize <- 20
A <- matrix(sample(seq_len(msize), msize^2, replace=TRUE), ncol=msize)
find_nearest_mode2D(A, 10, 1)
find_nearest_mode2D(A, 2, 4)
}
