% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_nearest_scaled_mode}
\alias{find_nearest_scaled_mode}
\title{Find the nearest maximum in a scaled array}
\usage{
find_nearest_scaled_mode(A, scale, ...)
}
\arguments{
\item{A}{An array.}

\item{scale}{A vector giving the value for one unit of distance. The length of \code{state} should be the same as the number of dimensions in \code{A}.}

\item{...}{The coordinates of a point within the dimensions of the array.}
}
\value{
Returns a list of the array positions of the nearest maximum that was found.
}
\description{
Find the nearest maximum in an array to a given point by hillclimbing accounting for different distances in each dimension.
}
\details{
A hillclimbing search from the input point across the array to find the nearest local maximum. 
The search is through each unit n-cube surrounding the next input point in the search (accounting for the array boundaries).
Differences in the scaled distance along each dimension are accounted for by finding the next maximum which is scaled by the distance e.g. (current_max - neighbour_max)/distance,
and select the value with the largest gradient. Ties are broken by random sampling. To make this reproducible the seed should be set.
}
\examples{
set.seed(444)
msize <- 20
A <- array(sample(seq_len(msize), msize^3, replace=TRUE), dim = rep(msize, 3))
find_nearest_scaled_mode(A, c(1,1,1), 11, 5, 5)
}
