% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_dist}
\alias{my_dist}
\title{Calculates distance between two matrices with the same number of columns.}
\usage{
my_dist(m1, m2)
}
\arguments{
\item{m1, m2}{Two matrices with the same number of columns.}
}
\value{
Returns a matrix of distances with dimension nrow(m1) by nrow(m2).
}
\description{
Calculates distance between two matrices with the same number of columns.
}
\details{
The distance is calculated between each row of the two input matrices.
}
\examples{
my_dist(diag(2), diag(1, 3, 2))
}
