% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normal_kernel_3d_indicator_vectorcpp_old}
\alias{normal_kernel_3d_indicator_vectorcpp_old}
\title{Calculates 3D Gaussian intensity over a grid -- OLD FUNCTION}
\usage{
normal_kernel_3d_indicator_vectorcpp_old(N, x, y, z, sd = 0.1)
}
\arguments{
\item{N}{An integer that defines where the intensity will be calculated in the unit cube into. If N=10 then 10^3 cubes are formed.}

\item{x, y, z}{A vector of positions on one dimension to calculate the kernel on.}

\item{sd}{The standard deviation which is assumed to be the same for each dimension.}
}
\value{
Returns a vector giving the intensity at each grid point.
}
\description{
Calculates 3D Gaussian intensity over a grid -- OLD FUNCTION
}
\examples{
set.seed(1)
N = 2
x = rnorm(5)
y = rnorm(5)
z = rnorm(5)
sd=1
normal_kernel_3d_indicator_vectorcpp(N, x, y, z, sd)
}
